Fix: Search page receives cookies and verify them
Currently after logged-in user goes to search page, the username is lost. This is confusing since users would not know if they logged in or not. The solution is redirect home page request to /v1/ path and make sure the search page verifies JWT token and use username from cookies to render.