Guard against user saving exactly same travel plans
Although the save buttons are disabled after clicking once, users would be able to save the same plans after reloading the page. This backend change prevent such behavior. In the future when we allow users to edit the plans, the plan views would get a new ID to differentiate from the original plan.