Feature: use precise location for travel planning
The trickier part for this backend change is how to use cache properly. Redis is using lat/lng for searching places. Therefore we need to feed the Redis client with precise location. At the same time, we need to maintain data freshness, therefore we need to reverse geocoding in order to find the city-level info in order to determine when to perform a new search.