Include price level in maps last search time
Update place freshness logic. Previously, we determine if a new API call is necessary when the number of results is lower than requested. However, with the new Maps query fixed on min/max price level, this will not work properly since higher priced restraunts are inevitably less than the requested most of the time. Therefore the new logic says as long as we have some data for the location with the requested category and price level, it is acceptable data from the database.