Experiement: use a new formula for places price level zero
This change affects all the places with Visit category, since all their current price levels are zero. Previously the place rating to pricing ratios for Visit places are treated flatly, and this affect the result of score since the final score will only depend on the number of user ratings. With this change, the ratings for Visit places will be taken into account.